+++ CHANGELOG.md
@@ -10,7 +10,11 @@ The project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html
 
 *Release Date: -**
 
+- Non-backward compatible changes:
+    - Hexital and Indicator 'append' will now only append to Candles and not sort placement.
+
 - Reworked and optimised 'append' and added 'insert' and 'prepend'
+- Added IndicatorCollection as a way to better reference Indicator's
 
 ---